refactor: split codegen.c into pipeline and per-node emission files
codegen.c was ~3,600 lines mixing two concerns: per-node C emission
(emit_expr, emit_stmt, emit_decl, RC helpers, type helpers) and
pipeline orchestration (codegen_emit phases, tuple collection, generic
instantiation). Split into codegen.c (~1,700 lines, pipeline) and
codegen_emit.c (~1,900 lines, per-node emission) with codegen_emit.h
as the internal interface.
